Add files via upload
This website showcases the craps games that an intergalactic species known as zerz plays. They use this game for their economy and as their monetary system. If one player loses, they are shot directly into the sun and the winner gets to move onto the next opponent. They use super advanced AI to randomize the outcome of the dice being rolled to make sure the game is as unpredictable as possible.